Create leo/unittest and its contents, including unittest/commands/editCommands.py
